% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/LST_MWA.R
\name{tau}
\alias{tau}
\title{Atmospheric transmittance calculation}
\usage{
tau(To = To, RH = To)
}
\arguments{
\item{To}{Near-surface air temperature (°C) of the date when Landsat passed over the study area}

\item{RH}{relative humidity (\%) of the date when Landsat passed over the study area}
}
\value{
Atmospheric transmittance
}
\description{
This fumction calculates Atmospheric transmittance from near-surface air temperature (To, °C) and relative humidity (RH, \%) of the date when Landsat passed over the study area
}
\examples{
tau(To = 26, RH = 42)
}
