% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/LST.R
\name{E_Valor}
\alias{E_Valor}
\title{Land Surface Emissivity according to Valor and Caselles 1996}
\usage{
E_Valor(NDVI)
}
\arguments{
\item{NDVI}{Raster* object, NDVI calculated from remote sensing imagery}
}
\value{
RasterLayer
}
\description{
This function calculates Land Surface Emissivity according to Valor and Caselles 1996
}
\examples{
NDVI <- raster::raster(ncol=100, nrow=100)
set.seed(2)
raster::values(NDVI) = runif(10000, min=0.02, max=0.8)
E_Valor(NDVI)
}
